requirements.txt
aiogram==3.17.0
SQLAlchemy==2.0.35
python-dotenv==1.0.1
aiosqlite==0.19.0
openai>=1.30.0
